refactor: view 패키지 제거
