+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CourseItemListener.kt
@@ -1,0 +1,7 @@
+package io.coursepick.coursepick.presentation.course
+
+interface CourseItemListener {
+    fun select(course: CourseItem)
+
+    fun navigateToMap(course: CourseItem)
+}
