+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesAction.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation.view.course
+package io.coursepick.coursepick.presentation.course
 
 import android.view.MenuItem
