+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesUiState.kt
@@ -1,6 +1,4 @@
-package io.coursepick.coursepick.presentation.view.course
-
-import io.coursepick.coursepick.presentation.model.course.CourseItem
+package io.coursepick.coursepick.presentation.course
 
 data class CoursesUiState(
     val courses: List<CourseItem>,
