+++ android/app/src/main/java/io/coursepick/coursepick/presentation/map/kakao/KakaoMapCameraController.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation.view.map.kakao
+package io.coursepick.coursepick.presentation.map.kakao
 
 import android.content.Context
 import android.location.Location
@@ -10,7 +10,7 @@ import com.kakao.vectormap.camera.CameraUpdateFactory
 import io.coursepick.coursepick.R
 import io.coursepick.coursepick.domain.Coordinate
 import io.coursepick.coursepick.domain.Segment
-import io.coursepick.coursepick.presentation.model.course.CourseItem
+import io.coursepick.coursepick.presentation.course.CourseItem
 
 class KakaoMapCameraController(
     context: Context,
