+++ android/app/src/main/java/io/coursepick/coursepick/presentation/routetfinder/OnChosenListener.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation.view.routetfinder
+package io.coursepick.coursepick.presentation.routetfinder
 
 interface OnChosenListener {
     fun onChosen(position: Int)
