+++ android/app/src/main/res/layout/activity_preferences.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/preferencesContainer"
-    android:name="io.coursepick.coursepick.presentation.PreferencesFragment"
+    android:name="io.coursepick.coursepick.presentation.view.preference.PreferencesFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
